fix: restore import paths and extract terminal guards after rebase (mcp/tool.py + terminal allowlist)
